+++ src/query/processor/cypher/astbuilder/ASTInternalNode.cpp
@@ -1,5 +1,6 @@
-//
-// Created by thamindu on 7/25/24.
-//
-
 #include "ASTInternalNode.h"
+
+void ASTInternalNode::addElements(ASTNode* element)
+{
+    elements.push_back(element);
+}
