+++ src/query/processor/cypher/astbuilder/ASTInternalNode.h
@@ -1,7 +1,4 @@
-//
-// Created by thamindu on 7/25/24.
-//
-#include "ASTStructure.h"
+#include "ASTNode.h"
 
 #ifndef ASTINTERNALNODE_H
 #define ASTINTERNALNODE_H
@@ -11,19 +8,12 @@ using namespace std;
 class ASTInternalNode : public ASTNode{
 
     public:
-
         ASTInternalNode(string nodeType)
         {
             this->nodeType = nodeType;
         }
 
-        void addElements(ASTNode* element) {
-            elements.push_back(element);
-        }
-
-
+        void addElements(ASTNode* element) ;
 };
-
-
 
 #endif //ASTINTERNALNODE_H
