+++ src/query/processor/cypher/astbuilder/ASTLeafNoValue.cpp
@@ -1,5 +1,1 @@
-//
-// Created by thamindu on 7/25/24.
-//
-
 #include "ASTLeafNoValue.h"
