+++ src/query/processor/cypher/astbuilder/ASTLeafNoValue.h
@@ -1,11 +1,6 @@
-//
-// Created by thamindu on 7/25/24.
-//
-#include "ASTStructure.h"
+#include "ASTNode.h"
 #ifndef ASTLEAFNOVALUE_H
 #define ASTLEAFNOVALUE_H
-
-
 
 class ASTLeafNoValue : public ASTNode {
 public:
