+++ src/query/processor/cypher/astbuilder/ASTLeafValue.cpp
@@ -1,5 +1,1 @@
-//
-// Created by thamindu on 7/25/24.
-//
-
 #include "ASTLeafValue.h"
