--- src/query/processor/cypher/astbuilder/ASTStructure.cpp
@@ -1,5 +1,0 @@
-//
-// Created by thamindu on 7/24/24.
-//
-
-#include "ASTStructure.h"
